requirements.txt
alembic==1.2.1
altgraph==0.16.1
appdirs==1.4.4
backpack==0.1
beautifulsoup4==4.8.1
blinker==1.4
cefpython3==66.0
cleo==0.6.8
cssutils==1.0.2
dataset==1.1.2
diff-match-patch==20181111
Faker==0.8.18
HTMLArk==1.0.0
inflection==0.3.1
jieba==0.39
lazy-object-proxy==1.4.2
lxml==4.4.2
macholib==1.11
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
modulegraph==0.17
numpy==1.17.2
ObjectListView==1.3.1
orator==0.9.9
pastel==0.1.1
peewee==3.14.4
pendulum==1.5.1
Pillow==6.2.2
py2app==0.19
pyaml==16.12.2
Pygments==2.4.2
PyInstaller==3.5
pylev==1.3.0
PyMySQL==0.9.3
Pypubsub==4.0.3
python-dateutil==2.8.0
python-editor==1.0.4
python-i18n==0.3.9
pytz==2019.2
pytzdata==2019.3
PyYAML==5.1.2
simplejson==3.16.0
six==1.12.0
soupsieve==1.9.4
SQLAlchemy==1.3.8
text-unidecode==1.2
tzlocal==1.5.1
Whoosh==2.7.4
wrapt==1.11.1
wxPython==4.1.1
